Refactored the folder structure
