+++ src/store/actions/index.js
@@ -1,0 +1,2 @@
+export { addIngrediant, removeIngrediant } from "./burgerBuilder";
+export {} from "./order";
